+++ exporters/trace/ocagent/src/main/java/io/opencensus/exporter/trace/ocagent/OcAgentTraceExporter.java
@@ -79,7 +79,8 @@ public final class OcAgentTraceExporter {
               configuration.getUseInsecure(),
               configuration.getSslContext(),
               configuration.getRetryInterval(),
-              configuration.getEnableConfig());
+              configuration.getEnableConfig(),
+              configuration.getDeadline());
       registerInternal(newHandler);
     }
   }